refactor: Add firebase dependency and update user model
This commit adds the firebase dependency to the project by updating the package.json file. It also modifies the user model to include new fields for followersCount and followingCount. These fields will be used to keep track of the number of followers and following for each user. This change enhances the user experience by providing more information about user relationships.

Fixes #N/A