refactor: Populate additional fields in getCourses and getMyProfile endpoints
This commit updates the `course.controller.js` and `user.controller.js` files to populate additional fields in the `getCourses` and `getMyProfile` endpoints. In the `getCourses` endpoint, the `madeBy` field is now populated with the `name` and `username` of the user who created the course. In the `getMyProfile` endpoint, the `courses` field is populated with the `madeBy` field, including the `name` and `username` of the user who created each course. This change enhances the user experience by providing more information about the course creators.

Fixes #N/A