feat: Add bookmarkCourse endpoint
This commit adds a new endpoint `bookmarkCourse` to the `course.controller.js` file. The endpoint allows users to bookmark a course by adding it to their list of bookmarked courses. It checks if the user and course exist, and if the course is already bookmarked by the user. If not, it adds the course to the user's bookmarkedCourses array and saves the changes. This functionality enhances the user experience by providing a way to easily bookmark courses.

Fixes #N/A