refactor: Add video upload support to createPost endpoint
This commit adds support for uploading videos to the createPost endpoint in the post.controller.js file. It introduces a new field, videoUrlBase64, in the request body and utilizes the uploadVideoOnCloudinaryBase64 function from the cloudinary.util.js file to upload the video to Cloudinary. This change enhances the functionality of the application by allowing users to include videos in their posts.

Fixes #N/A