feat: Add removeFollower endpoint to user controller
This commit adds the removeFollower endpoint to the user.controller.js file. It allows users to remove a follower from their followers list. The endpoint takes the follower's ID as a parameter and updates the user's followers and following counts accordingly. This change enhances the functionality of the application by providing users with the ability to manage their followers.

Fixes #N/A